+ 바닐라 무기 정보 읽어오는 기능 구현 + 무기 리스트 메인 화면에 추가(UI 개선 예정됨) = 기본 바닐라 무기 정보 표시 패널 UI 작업 진행중 + 무기 선택 리스트에서 정보를 표시할 무기 선택 기능 추가

+++ Source/USECEAPG_Main/USECEAPG_Main/Logic/VanillaCompatProjectile.cs
@@ -11,9 +11,9 @@ namespace USECEAPG_Main.Logic
         public string DefName { get; set; }
         public string Label { get; set; }
         public string DamageDef { get; set; }
-        public int DamageAmountBase { get; set; } 
-        public int StoppingPower { get; set; }
-        public int Speed { get; set; }
+        public double DamageAmountBase { get; set; } 
+        public double StoppingPower { get; set; }
+        public double Speed { get; set; }
         public VanillaCompatGraphicData Graphic { get; set; }
     }
 }
